ViewLogStaff_ApproveTrashEvent

+++ GreenCorner.MVC/Controllers/ProductController.cs
@@ -3,17 +3,20 @@ using GreenCorner.MVC.Services.Interface;
 using GreenCorner.MVC.Utility;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace GreenCorner.MVC.Controllers
 {
     public class ProductController : Controller
     {
         private readonly IProductService _productService;
-        public ProductController(IProductService productService)
-        {
-            _productService = productService;
-        }
-        public async Task<IActionResult> Index()
+        private readonly IAdminService _adminService;
+		public ProductController(IProductService productService, IAdminService adminService)
+		{
+			_productService = productService;
+			_adminService = adminService;
+		}
+		public async Task<IActionResult> Index()
         {
             List<ProductDTO> listProduct = new();
             ResponseDTO? response = await _productService.GetAllProduct();
@@ -37,7 +40,13 @@ namespace GreenCorner.MVC.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateNewProduct(ProductDTO productDTO)
         {
-            var files = Request.Form.Files;
+			if (!User.Identity.IsAuthenticated)
+			{
+				TempData["loginError"] = "Bạn cần đăng nhập để thực hiện được chức năng này";
+				return RedirectToAction("Login", "Auth");
+			}
+
+			var files = Request.Form.Files;
 
             var (isSuccess, imagePaths, errorMessage) = await FileUploadHelper.UploadImagesStrictAsync(
                 files, folderName: "product", filePrefix: "product");
@@ -56,7 +65,19 @@ namespace GreenCorner.MVC.Controllers
                 if (response != null && response.IsSuccess)
                 {
                     TempData["success"] = "Product created successfully!";
-                    return RedirectToAction(nameof(Index));
+					ProductDTO product = JsonConvert.DeserializeObject<ProductDTO>(response.Result.ToString());
+                    var StaffName = User.Claims.Where(u => u.Type == JwtRegisteredClaimNames.Name).FirstOrDefault()?.Value;
+                    var log = new SystemLogDTO()
+                    {
+                        UserId = User.Claims.Where(u => u.Type == JwtRegisteredClaimNames.Sub).FirstOrDefault()?.Value,
+                        ActionType = "Thêm mới",
+                        ObjectType = "Sản phẩm",
+                        ObjectId = product.ProductId,
+                        Description = $"Nhân viên {StaffName} đã thêm mới sản phẩm {product.Name} với ID {product.ProductId}",
+                        CreatedAt = DateTime.Now,
+                    };
+					var logResponse = await _adminService.AddLogStaff(log);
+					return RedirectToAction(nameof(Index));
                 }
 
                 TempData["error"] = response?.Message ?? "Có lỗi xảy ra.";
@@ -84,11 +105,27 @@ namespace GreenCorner.MVC.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteProduct(ProductDTO productDto)
         {
-            ResponseDTO response = await _productService.DeleteProduct(productDto.ProductId);
+			if (!User.Identity.IsAuthenticated)
+			{
+				TempData["loginError"] = "Bạn cần đăng nhập để thực hiện được chức năng này";
+				return RedirectToAction("Login", "Auth");
+			}
+			ResponseDTO response = await _productService.DeleteProduct(productDto.ProductId);
             if (response != null && response.IsSuccess)
             {
-                TempData["success"] = "Product deleted successfully!";
-                return RedirectToAction(nameof(Index));
+                TempData["success"] = "Xóa sản phẩm thành công!";
+				var StaffName = User.Claims.Where(u => u.Type == JwtRegisteredClaimNames.Name).FirstOrDefault()?.Value;
+				var log = new SystemLogDTO()
+				{
+					UserId = User.Claims.Where(u => u.Type == JwtRegisteredClaimNames.Sub).FirstOrDefault()?.Value,
+					ActionType = "Xóa",
+					ObjectType = "Sản phẩm",
+					ObjectId = productDto.ProductId,
+					Description = $"Nhân viên {StaffName} đã thêm mới sản phẩm {productDto.Name} với ID {productDto.ProductId}",
+					CreatedAt = DateTime.Now,
+				};
+				var logResponse = await _adminService.AddLogStaff(log);
+				return RedirectToAction(nameof(Index));
             }
             else
             {
@@ -115,7 +152,13 @@ namespace GreenCorner.MVC.Controllers
         [HttpPost]
         public async Task<IActionResult> UpdateProduct(ProductDTO productDto)
         {
-            var files = Request.Form.Files;
+			if (!User.Identity.IsAuthenticated)
+			{
+				TempData["loginError"] = "Bạn cần đăng nhập để thực hiện được chức năng này";
+				return RedirectToAction("Login", "Auth");
+			}
+
+			var files = Request.Form.Files;
 
             bool hasNewImages = files != null && files.Count > 0;
 
@@ -148,8 +191,19 @@ namespace GreenCorner.MVC.Controllers
             var response = await _productService.UpdateProduct(productDto);
             if (response != null && response.IsSuccess)
             {
-                TempData["success"] = "Product updated successfully!";
-                return RedirectToAction(nameof(Index));
+                TempData["success"] = "Cập nhật sản phẩm thành công!";
+				var StaffName = User.Claims.Where(u => u.Type == JwtRegisteredClaimNames.Name).FirstOrDefault()?.Value;
+				var log = new SystemLogDTO()
+				{
+					UserId = User.Claims.Where(u => u.Type == JwtRegisteredClaimNames.Sub).FirstOrDefault()?.Value,
+					ActionType = "Cập nhật",
+					ObjectType = "Sản phẩm",
+					ObjectId = productDto.ProductId,
+					Description = $"Nhân viên {StaffName} đã thêm mới sản phẩm {productDto.Name} với ID {productDto.ProductId}",
+					CreatedAt = DateTime.Now,
+				};
+				var logResponse = await _adminService.AddLogStaff(log);
+				return RedirectToAction(nameof(Index));
             }
 
             TempData["error"] = response?.Message;
